Allow the models to be extended.
This allows the metadata specification to be expanded on within some
platforms where additional elements can be well defined (and not just
properties).
